requirements.txt
PyMySQL~=1.1.0
customtkinter~=5.2.1
future~=0.18.3
pillow~=10.2.0